Move host and hello fragment to kotlin folder
