+++ app/src/main/kotlin/com/backbase/golden_sample_app/presentation/HelloFragment.kt
@@ -1,10 +1,11 @@
-package com.backbase.golden_sample_app
+package com.backbase.golden_sample_app.presentation
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.backbase.golden_sample_app.R
 
 /**
  * A simple [Fragment] subclass.
