+++ app/src/main/res/layout/fragment_app_nav_host.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AppNavHostFragment">
+    tools:context=".host.AppNavHostFragment">
 
     <TextView
         android:layout_width="match_parent"
